+++ pmd_beamphysics/fields/corrector_modeling.py
@@ -3,7 +3,8 @@ from matplotlib import pyplot as plt
 from scipy.constants import mu_0 as u0
 from scipy.constants import pi
 
-from pmd_beamphysics import FieldMesh
+from .fieldmesh import FieldMesh
+from ..utils import get_rotation_matrix
 
 
 def set_axes_equal(ax):
@@ -167,51 +168,6 @@ def bfield_from_thin_straight_wire(
 
     else:
         return B[:, :, :, 0], B[:, :, :, 1], B[:, :, :, 2]
-
-
-def Ry(angle):
-    """Defines a roation in around +y direction"""
-
-    C = np.cos(angle)
-    S = np.sin(angle)
-
-    # Fast, manual construction of the matrix
-    return np.array([[C, 0, S], [0, 1, 0], [-S, 0, C]])
-
-
-def Rx(angle):
-    """Defines a roation around +x direction"""
-
-    C = np.cos(angle)
-    S = np.sin(angle)
-
-    return np.array([[1, 0, 0], [0, C, -S], [0, +S, C]])
-
-
-def Rz(angle):
-    """Defines a roation arounx +z direction"""
-
-    C = np.cos(angle)
-    S = np.sin(angle)
-
-    return np.array([[C, -S, 0], [S, C, 0], [0, 0, 1]])
-
-
-def get_rotation_matrix(pitch, yaw, tilt):
-    """Defines a general 3d rotation in terms of the orientation angles theta, phi, psi"""
-
-    if pitch is None:
-        pitch = 0
-    if yaw is None:
-        yaw = 0
-    if tilt is None:
-        tilt = 0
-
-    ry = Ry(pitch)
-    rx = Rx(yaw)
-    rz = Rz(tilt)
-
-    return np.matmul(ry, np.matmul(rx, rz))
 
 
 def bfield_from_thin_rectangular_coil(
